pyproject.toml
[project]
name = "lab-dispatcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-cloud-pubsub>=2.31.0",
]
